Renamed original AlgorithmTimer class
